Replace build-system with CMake
