+++ License.txt
@@ -1,0 +1,3 @@
+No idea, I didn't write this.
+
+So (c) whenever, whoever. All rights (probably) reserved.
