+++ README.md
@@ -1,0 +1,7 @@
+YOLO.
+
+## Requirements
+
+* CMake
+* C++ toolchain
+* TinyXML
