--- Swr.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26430.16
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "WxHexEditorColorMaker", "tools\WxHexEditorColorMaker\WxHexEditorColorMaker.vcxproj", "{D6827F75-A090-4BE8-BE49-F1AC027B26BE}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "XmlConverter", "tools\XmlConverter\XmlConverter.vcxproj", "{A67BA80F-93BA-4F40-8181-971002D70F17}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x86 = Debug|x86
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D6827F75-A090-4BE8-BE49-F1AC027B26BE}.Debug|x86.ActiveCfg = Debug|Win32
-		{D6827F75-A090-4BE8-BE49-F1AC027B26BE}.Debug|x86.Build.0 = Debug|Win32
-		{D6827F75-A090-4BE8-BE49-F1AC027B26BE}.Release|x86.ActiveCfg = Release|Win32
-		{D6827F75-A090-4BE8-BE49-F1AC027B26BE}.Release|x86.Build.0 = Release|Win32
-		{A67BA80F-93BA-4F40-8181-971002D70F17}.Debug|x86.ActiveCfg = Debug|Win32
-		{A67BA80F-93BA-4F40-8181-971002D70F17}.Debug|x86.Build.0 = Debug|Win32
-		{A67BA80F-93BA-4F40-8181-971002D70F17}.Release|x86.ActiveCfg = Release|Win32
-		{A67BA80F-93BA-4F40-8181-971002D70F17}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
