--- tools/WxHexEditorColorMaker/WxHexEditorColorMaker.vcxproj.filters
@@ -1,60 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <ItemGroup>
-    <Filter Include="Fichiers sources">
-      <UniqueIdentifier>{4FC737F1-C7A5-4376-A066-2A32D752A2FF}</UniqueIdentifier>
-      <Extensions>cpp;c;cc;cxx;def;odl;idl;hpj;bat;asm;asmx</Extensions>
-    </Filter>
-    <Filter Include="Fichiers d%27en-tête">
-      <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
-      <Extensions>h;hh;hpp;hxx;hm;inl;inc;xsd</Extensions>
-    </Filter>
-    <Filter Include="Fichiers de ressources">
-      <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
-      <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
-    </Filter>
-    <Filter Include="Fichiers sources\Extern">
-      <UniqueIdentifier>{088361d2-6913-43fd-88bc-04b330af2fb3}</UniqueIdentifier>
-    </Filter>
-  </ItemGroup>
-  <ItemGroup>
-    <ClCompile Include="main.cpp">
-      <Filter>Fichiers sources</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\include\EMO_BaseFile.cpp">
-      <Filter>Fichiers sources</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\include\Swr_Model.cpp">
-      <Filter>Fichiers sources</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\include\common.cpp">
-      <Filter>Fichiers sources</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\depends\half\half.cpp">
-      <Filter>Fichiers sources\Extern</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\depends\tinyxml\tinystr.cpp">
-      <Filter>Fichiers sources\Extern</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\depends\tinyxml\tinyxml.cpp">
-      <Filter>Fichiers sources\Extern</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\depends\tinyxml\tinyxmlerror.cpp">
-      <Filter>Fichiers sources\Extern</Filter>
-    </ClCompile>
-    <ClCompile Include="..\..\depends\tinyxml\tinyxmlparser.cpp">
-      <Filter>Fichiers sources\Extern</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="..\..\include\Swr_Model.h">
-      <Filter>Fichiers sources</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\include\common.h">
-      <Filter>Fichiers sources</Filter>
-    </ClInclude>
-    <ClInclude Include="..\..\include\EMO_BaseFile.h">
-      <Filter>Fichiers sources</Filter>
-    </ClInclude>
-  </ItemGroup>
-</Project>
